+++ BE-Movie/src/main/java/com/a2m/service/ShowtimeService.java
@@ -13,7 +13,7 @@ public interface ShowtimeService {
 
     List<Movies> getMovieByDate(Date date);
 
-    List<Showtimes> getShowTimeByDate(Date date, int id);
+    List<Showtimes> getShowTimeByDate(Date date, Long id);
     
 //    List<Showtimes> getShowtimesByDate(String date);
 
@@ -24,5 +24,9 @@ public interface ShowtimeService {
     List<ShowtimeDateResponse> findByMovie(Long idMovie);
 
     List<Showtimes> getShowtimeByMovieAndDate(Long idMovie, Date date);
+
+//	List<Showtimes> getShowTimeByDate(Date date, Long id);
+
+//	List<ShowtimeResponse> today();
 
 }
